+++ Runtime/Tween/TweenPunchMode.cs
@@ -1,0 +1,9 @@
+namespace Dreamy.UI
+{
+    public enum TweenPunchMode
+    {
+        Position,
+        Rotation,
+        Scale
+    }
+}